final commit for working rating system

+++ src/models/Bookmark.rb
@@ -134,7 +134,7 @@ class Bookmark
       count = 0
       totalRatings = Rating.getByBookmarkId(bookmarkId)
       for j in (0..(totalRatings.length - 1))
-        totalRating += totalRatings[j].rating.to_i
+        totalRating += totalRatings[j].rating.to_f
         count += 1
       end
       if count == 0 then
